Try to put order in this caos-1
