+++ src/main/java/xarmanta/mainwindow/model/GitContext.kt
@@ -1,4 +1,4 @@
-package xarmanta.mainwindow.shared
+package xarmanta.mainwindow.model
 
 import xarmanta.mainwindow.shared.exceptions.InvalidUrl
 import java.io.File
