+++ src/main/resources/main.fxml
@@ -21,7 +21,7 @@
 <?import javafx.scene.text.Font?>
 <?import javafx.scene.text.TextFlow?>
 
-<StackPane fx:id="root" prefHeight="600.0" prefWidth="1080.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="xarmanta.mainwindow.infraestructure.MainWindowController">
+<StackPane fx:id="root" prefHeight="600.0" prefWidth="1080.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="xarmanta.mainwindow.infraestructure.javafx.MainWindowController">
 <VBox fx:id="vBox">
   <children>
     <MenuBar VBox.vgrow="NEVER">
