+++ src/test/java/xarmant/XGitTest.kt
@@ -1,14 +1,13 @@
 package xarmant
 
 import org.eclipse.jgit.api.Git
-import org.junit.jupiter.api.AfterAll
 import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.TestInstance
 import xarmant.mainwindow.infraestructure.ConsoleMonitor
 import xarmanta.mainwindow.model.Commit
-import xarmanta.mainwindow.shared.GitContext
-import xarmanta.mainwindow.shared.XGit
+import xarmanta.mainwindow.model.GitContext
+import xarmanta.mainwindow.shared.git.XGit
 import java.io.File
 import java.nio.file.Files
 import java.nio.file.Path
